Note failure to save files, sometimes, in bug list at top.
This is a pain to debug, since it only fails for simulations that take a long time to run.  I added some drawnow commands in various places - maybe the problem is that saveas doesn't force a drawnow and when there is a lot going on the figure isn't ready when the saveas command executes.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>