Fixed information leak.  It was a little interesting.  Also set up to run through a large set of conditions on the cluster
The leak came because the set of cones being sampled for blank and test in the second site calculations were not the same.  Thus, the classifier could use differences in the noise instantiation (and also, perhaps, spatial non-uniformities in the responses to the uniform field) to do the classifying.  This turns out to be enough to allow good performance.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>