Trying to rationalize tree structure
The idea is that we can all put computationaleyebrain/toobox onto our paths.  I created three subdirectories, one for routines that seem likely to end up in isetbio eventually, one for routines that seem likely to end up in the psychophysics toolbox eventually, and one for routines that seem specific to this general project but more general than just one computation.